+++ Code/Tools/W3DView/AssetInfo.h
@@ -68,7 +68,7 @@ class AssetInfoClass
 			  m_dwUserData (0L),
 			  m_pRenderObj (NULL)			{ Initialize (); }
 
-		AssetInfoClass (LPCTSTR passet_name, ASSET_TYPE type, RenderObjClass *prender_obj = NULL, DWORD user_data = 0L)
+		AssetInfoClass (LPCTSTR passet_name, ASSET_TYPE type, RenderObjClass *prender_obj = NULL, uintptr_t user_data = 0L)
 			: m_Name (passet_name),
 			  m_AssetType (type),
 			  m_dwUserData (user_data),
@@ -88,14 +88,14 @@ class AssetInfoClass
 		const CString &	Get_Hierarchy_Name (void) const			{ return m_HierarchyName; }
 		const CString &	Get_Original_Name (void) const			{ return m_OriginalName; }
 		ASSET_TYPE			Get_Type (void) const						{ return m_AssetType; }
-		DWORD					Get_User_Number (void) const				{ return m_dwUserData; }
+		uintptr_t					Get_User_Number (void) const				{ return m_dwUserData; }
 		const CString &	Get_User_String (void) const				{ return m_UserString; }
 		RenderObjClass *	Get_Render_Obj (void) const				{ SAFE_ADD_REF (m_pRenderObj); return m_pRenderObj; }
 		RenderObjClass *	Peek_Render_Obj (void) const				{ return m_pRenderObj; }
 		void					Set_Name (LPCTSTR pname)					{ m_Name = pname; }
 		void					Set_Hierarchy_Name (LPCTSTR pname)		{ m_HierarchyName = pname; }
 		void					Set_Type (ASSET_TYPE type)					{ m_AssetType = type; }
-		void					Set_User_Number (DWORD user_data)		{ m_dwUserData = user_data; }
+		void					Set_User_Number (uintptr_t user_data)		{ m_dwUserData = user_data; }
 		void					Set_User_String (LPCTSTR string)			{ m_UserString = string; }
 		void					Set_Render_Obj (RenderObjClass *pobj)	{ MEMBER_ADD (m_pRenderObj, pobj); }
 
@@ -124,7 +124,7 @@ class AssetInfoClass
 		CString				m_UserString;
 		CString				m_OriginalName;
 		ASSET_TYPE			m_AssetType;
-		DWORD					m_dwUserData;
+		uintptr_t			m_dwUserData;
 		RenderObjClass *	m_pRenderObj;
 };
